backend/requirements.txt
ultralytics==8.0.196
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy>=1.26.0
opencv-python==4.8.1.78
torch>=2.0.0
torchvision>=0.15.0